hetzner-service/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
transformers
accelerate
python-dotenv
sentencepiece
pytest
evaluate
fastapi
uvicorn[standard]
pydantic
grpcio
grpcio-tools
protobuf
